requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0

# Diffusion models
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1

# Vision-Language models
openai>=1.0.0
timm>=0.9.0

# Data processing
opencv-python>=4.8.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Evaluation
pycocotools>=2.0.6

# Optional: For caption synthesis
# llama  # Install separately if needed
# llava  # Install separately if needed
